+++ tests/testthat/test-prune_occurrences.R
@@ -42,7 +42,7 @@ test_that("keep_rows keeps everything if condition is always `TRUE`", {
 
 test_that("keep_content_rows works as expected", {
   tab <- get_table()
-  more_than_twenty <- has_count_in_cols(above = 20L, col_names = names(tab))
+  more_than_twenty <- has_count_in_cols(atleast = 21L, col_names = names(tab))
   result <- prune_table(tab, keep_content_rows(more_than_twenty))
   result_leaves <- collect_leaves(result)
   result_content_rows <- result_leaves[sapply(result_leaves, class) == "ContentRow"]
@@ -54,7 +54,7 @@ test_that("keep_content_rows works as expected", {
 
 test_that("has_count_in_cols result works in a special case identical to standard pruning", {
   tab <- get_table()
-  row_condition <- has_count_in_cols(above = 0L, col_names = names(tab))
+  row_condition <- has_count_in_cols(atleast = 1L, col_names = names(tab))
   result <- prune_table(tab, keep_rows(row_condition))
   expected <- prune_table(tab)
   expect_identical(result, expected)
@@ -71,17 +71,36 @@ test_that("has_count_in_cols result performs comparisons correctly", {
     )
   )
   table_row <- collect_leaves(sub_tab)[[1]]
-  row_condition <- has_count_in_cols(above = 3L, col_names = "A: Drug X")
+  row_condition <- has_count_in_cols(atleast = 4L, col_names = "A: Drug X")
   expect_false(row_condition(table_row))
-  row_condition <- has_count_in_cols(above = 0L, col_names = "A: Drug X")
+  row_condition <- has_count_in_cols(atleast = 1L, col_names = "A: Drug X")
   expect_true(row_condition(table_row))
-  row_condition <- has_count_in_cols(above = 4L, col_names = c("A: Drug X", "B: Placebo"))
+  row_condition <- has_count_in_cols(atleast = 5L, col_names = c("A: Drug X", "B: Placebo"))
+  expect_true(row_condition(table_row))
+})
+
+test_that("has_count_in_any_col result performs comparisons correctly", {
+  tab <- get_table()
+  sub_tab <- tab[5, ]
+  expect_identical(
+    to_string_matrix(sub_tab),
+    structure(
+      c("", "BRA", "A: Drug X", "1 (3.7%)", "B: Placebo", "4 (20%)", "C: Combination", "1 (3.2%)"),
+      .Dim = c(2L, 4L)
+    )
+  )
+  table_row <- collect_leaves(sub_tab)[[1]]
+  row_condition <- has_count_in_any_col(atleast = 2L, col_names = "A: Drug X")
+  expect_false(row_condition(table_row))
+  row_condition <- has_count_in_any_col(atleast = 1L, col_names = "A: Drug X")
+  expect_true(row_condition(table_row))
+  row_condition <- has_count_in_any_col(atleast = 4L, col_names = c("A: Drug X", "B: Placebo"))
   expect_true(row_condition(table_row))
 })
 
 test_that("has_fraction_in_cols result works in a special case identical to standard pruning", {
   tab <- get_table()
-  row_condition <- has_fraction_in_cols(above = 0, col_names = names(tab))
+  row_condition <- has_fraction_in_cols(atleast = 0.000001, col_names = names(tab))
   result <- prune_table(tab, keep_rows(row_condition))
   expected <- prune_table(tab)
   expect_identical(result, expected)
@@ -102,11 +121,30 @@ test_that("has_fraction_in_cols result performs comparisons correctly", {
     c(121L, 106L, 129L)
   )
   table_row <- collect_leaves(sub_tab)[[1]]
-  row_condition <- has_fraction_in_cols(above = 0.01, col_names = "A: Drug X")
+  row_condition <- has_fraction_in_cols(atleast = 0.01, col_names = "A: Drug X")
   expect_false(row_condition(table_row))
-  row_condition <- has_fraction_in_cols(above = 0.008, col_names = "A: Drug X")
+  row_condition <- has_fraction_in_cols(atleast = 0.008, col_names = "A: Drug X")
   expect_true(row_condition(table_row))
-  row_condition <- has_fraction_in_cols(above = 0.02, col_names = c("A: Drug X", "B: Placebo"))
+  row_condition <- has_fraction_in_cols(atleast = 0.02, col_names = c("A: Drug X", "B: Placebo"))
+  expect_true(row_condition(table_row))
+})
+
+test_that("has_fraction_in_any_col result performs comparisons correctly", {
+  tab <- get_table()
+  sub_tab <- tab[5, ]
+  expect_identical(
+    to_string_matrix(sub_tab),
+    structure(
+      c("", "BRA", "A: Drug X", "1 (3.7%)", "B: Placebo", "4 (20%)", "C: Combination", "1 (3.2%)"),
+      .Dim = c(2L, 4L)
+    )
+  )
+  table_row <- collect_leaves(sub_tab)[[1]]
+  row_condition <- has_fraction_in_any_col(atleast = 0.038, col_names = c("A: Drug X", "C: Combination"))
+  expect_false(row_condition(table_row))
+  row_condition <- has_fraction_in_any_col(atleast = 0.036, col_names = c("A: Drug X", "C: Combination"))
+  expect_true(row_condition(table_row))
+  row_condition <- has_fraction_in_any_col(atleast = 0.2, col_names = c("A: Drug X", "B: Placebo"))
   expect_true(row_condition(table_row))
 })
 
@@ -121,11 +159,11 @@ test_that("has_fractions_difference result performs comparisons correctly", {
     )
   )
   table_row <- collect_leaves(sub_tab)[[1]]
-  row_condition <- has_fractions_difference(above = 0.01, col_names = c("A: Drug X", "B: Placebo"))
+  row_condition <- has_fractions_difference(atleast = 0.01, col_names = c("A: Drug X", "B: Placebo"))
   expect_true(row_condition(table_row))
-  row_condition <- has_fractions_difference(above = 0.5, col_names = c("A: Drug X", "B: Placebo"))
+  row_condition <- has_fractions_difference(atleast = 0.5, col_names = c("A: Drug X", "B: Placebo"))
   expect_false(row_condition(table_row))
-  row_condition <- has_fractions_difference(above = 0.01, col_names = c("A: Drug X", "C: Combination"))
+  row_condition <- has_fractions_difference(atleast = 0.01, col_names = c("A: Drug X", "C: Combination"))
   expect_false(row_condition(table_row))
 })
 
@@ -140,11 +178,11 @@ test_that("has_counts_difference result performs comparisons correctly", {
     )
   )
   table_row <- collect_leaves(sub_tab)[[1]]
-  row_condition <- has_counts_difference(above = 2L, col_names = c("A: Drug X", "B: Placebo"))
+  row_condition <- has_counts_difference(atleast = 3L, col_names = c("A: Drug X", "B: Placebo"))
   expect_true(row_condition(table_row))
-  row_condition <- has_counts_difference(above = 3L, col_names = c("A: Drug X", "B: Placebo"))
+  row_condition <- has_counts_difference(atleast = 4L, col_names = c("A: Drug X", "B: Placebo"))
   expect_false(row_condition(table_row))
-  row_condition <- has_counts_difference(above = 0L, col_names = c("A: Drug X", "C: Combination"))
+  row_condition <- has_counts_difference(atleast = 1L, col_names = c("A: Drug X", "C: Combination"))
   expect_false(row_condition(table_row))
 })
 
@@ -153,9 +191,9 @@ test_that("combination of pruning functions works", {
   result <- tab %>%
     prune_table(
       keep_rows(
-        has_fractions_difference(above = 0.1, c("A: Drug X", "C: Combination")) &
-          has_counts_difference(above = 1L, c("B: Placebo", "A: Drug X")) &
-          has_count_in_cols(above = 2L, "A: Drug X")
+        has_fractions_difference(atleast = 0.1, c("A: Drug X", "C: Combination")) &
+          has_counts_difference(atleast = 2L, c("B: Placebo", "A: Drug X")) &
+          has_count_in_cols(atleast = 3L, "A: Drug X")
       )
     )
   result_matrix <- to_string_matrix(result)