+++ tests/testthat/test-score_occurrences.R
@@ -1,19 +1,26 @@
-test_that("score_occurrences functions as expected with valid input and default arguments", {
-  library(dplyr)
+library(dplyr)
+
+get_df_ae <- function() {
   set.seed(1)
   dfsl <- data.frame(
     USUBJID = as.character(c(1, 2, 3, 4, 5)),
     ARM = sample(c("A", "B", "C"), 5, replace = TRUE)
   )
-  N_per_arm <- table(dfsl$ARM) # nolint
 
   dfae <- data.frame(
     USUBJID = as.character(c(1, 2, 3, 4)),
     AEBODSYS = sample(c("AEBS1", "AEBS2"), 20, replace = TRUE),
     AEDECOD = sample(c("AEPT1", "AEPT2", "AEPT3"), 20, replace = TRUE)
   )
-  dfae <- dfae %>% arrange(USUBJID, AEBODSYS, AEDECOD)
+  dfae <- dfae %>% arrange(USUBJID, AEBODSYS, AEDECOD) #nolint
   dfae <- left_join(dfae, dfsl, by = "USUBJID")
+  structure(
+    dfae,
+    n_per_arm = table(dfsl$ARM)
+  )
+}
+
+get_full_table <- function() {
 
   lyt <- basic_table() %>%
     split_cols_by("ARM") %>%
@@ -35,13 +42,19 @@ test_that("score_occurrences functions as expected with valid input and default 
       )) %>%
     count_occurrences(vars = "AEDECOD")
 
-  rtable_object <- build_table(lyt, dfae, col_counts = N_per_arm) %>%
-    prune_table()
+  dfae <- get_df_ae()  #nolint
 
-  rtable_object_sorted <- rtable_object %>%
+  build_table(lyt, dfae, col_counts = attr(dfae, "n_per_arm")) %>%
+    prune_table()
+}
+
+test_that("score_occurrences functions as expected", {
+  full_table <- get_full_table()
+
+  sorted_table <- full_table %>%
     sort_at_path(path =  c("AEBODSYS", "*", "AEDECOD"), scorefun = score_occurrences)
 
-  result <- to_string_matrix(rtable_object_sorted)
+  result <- to_string_matrix(sorted_table)
 
   expected <- rbind(
     c("",    "A",         "B",        "C"),
@@ -63,4 +76,63 @@ test_that("score_occurrences functions as expected with valid input and default 
   )
 
   expect_equal(result, expected)
+})
+
+test_that("score_occurrences_cols functions as expected", {
+  full_table <- get_full_table()
+
+  score_col_c <- score_occurrences_cols(col_names = "C")
+  expect_is(score_col_c, "function")
+
+  sorted_table <- full_table %>%
+    sort_at_path(path =  c("AEBODSYS", "*", "AEDECOD"), scorefun = score_col_c)
+
+  result <- to_string_matrix(sorted_table)
+
+  expected <- structure(
+    c("", "", "Total number of patients with at least one event",
+      "Total number of events", "AEBS1", "Total number of patients with at least one event",
+      "Total number of events", "AEPT1", "AEPT2", "AEPT3", "AEBS2",
+      "Total number of patients with at least one event", "Total number of events",
+      "AEPT3", "AEPT1", "AEPT2", "A", "(N=3)", "2 (66.7%)", "10", "",
+      "2 (66.7%)", "7", "2 (66.7%)", "2 (66.7%)", "1 (33.3%)", "",
+      "2 (66.7%)", "3", "0", "1 (33.3%)", "2 (66.7%)", "B", "(N=1)",
+      "1 (100%)", "5", "", "1 (100%)", "3", "1 (100%)", "1 (100%)",
+      "1 (100%)", "", "1 (100%)", "2", "0", "0", "1 (100%)", "C", "(N=1)",
+      "1 (100%)", "5", "", "1 (100%)", "4", "1 (100%)", "1 (100%)",
+      "1 (100%)", "", "1 (100%)", "1", "1 (100%)", "0", "0"),
+    .Dim = c(16L, 4L)
+  )
+  expect_identical(result, expected)
+})
+
+test_that("score_occurrences_subtable functions as expected", {
+  dfae <- get_df_ae()
+
+  full_table <- basic_table() %>%
+    split_cols_by("ARM") %>%
+    add_colcounts() %>%
+    split_rows_by("AEBODSYS", child_labels = "visible", nested = FALSE)  %>%
+    count_occurrences(vars = "AEDECOD") %>%
+    build_table(dfae, col_counts = attr(dfae, "n_per_arm")) %>%
+    prune_table()
+
+  score_subtable_all <- score_occurrences_subtable(col_names = names(full_table))
+  expect_is(score_subtable_all, "function")
+
+  sorted_table <- full_table %>%
+    sort_at_path(path =  c("AEBODSYS"), scorefun = score_subtable_all, decreasing = FALSE)
+
+  result <- to_string_matrix(sorted_table)
+
+  expected <- structure(
+    c("", "", "AEBS2", "AEPT1", "AEPT2", "AEPT3", "AEBS1",
+      "AEPT1", "AEPT2", "AEPT3", "A", "(N=3)", "", "1 (33.3%)", "2 (66.7%)",
+      "0", "", "2 (66.7%)", "2 (66.7%)", "1 (33.3%)", "B", "(N=1)",
+      "", "0", "1 (100%)", "0", "", "1 (100%)", "1 (100%)", "1 (100%)",
+      "C", "(N=1)", "", "0", "0", "1 (100%)", "", "1 (100%)", "1 (100%)",
+      "1 (100%)"),
+    .Dim = c(10L, 4L)
+  )
+  expect_identical(result, expected)
 })