+++ tests/testthat/test-utils.R
@@ -24,6 +24,13 @@ test_that("f_conf_level fails for non-proportion input", {
   expect_error(f_conf_level(-1))
 })
 
+test_that("make_names works as expected", {
+  nams <- c("Any Grade (%)", "Total AE numbers!", "No adverse events ...")
+  result <- make_names(nams)
+  expected <- c("AnyGrade", "TotalAEnumbers", "Noadverseevents")
+  expect_identical(result, expected)
+})
+
 test_that("get_covariates works for a character vector", {
   result <- get_covariates(c("a * b", "c"))
   expected <- list(a = "a", b = "b", c = "c")